\name{as.array.chain}
\alias{as.array.chain}
\title{Transform a chain to an array}
\usage{
  \method{as.array}{chain} (x, ...)
}
\arguments{
  \item{x}{chain object}

  \item{...}{passed to simplify2array}
}
\value{
  an array (or possibly a matrix, or vector if appropriate)
}
\description{
  Not much else to say about this.
}
\author{
  Grady Weyenberg
}

