\name{as.mcmc.list.chain}
\alias{as.mcmc.list.chain}
\title{Convert a chain object to a coda mcmc.list object}
\usage{
  \method{as.mcmc.list}{chain} (x, ...)
}
\arguments{
  \item{x}{chain to be converted}

  \item{...}{unused arguments}
}
\value{
  a coda mcmc.list object
}
\description{
  Convert a chain object to a coda mcmc.list object
}
\author{
  Grady Weyenberg
}

