\name{crossover}
\alias{crossover}
\title{Create a crossover updater}
\usage{
  crossover(target, temps)
}
\arguments{
  \item{target}{target density}

  \item{temps}{temperature ladder}
}
\value{
  a function that implements the crossover updater
}
\description{
  Create a crossover updater
}
\author{
  Grady Weyenberg
}

