\name{[.chain}
\alias{[.chain}
\alias{[<-.chain}
\title{Need this to keep the extraction operator from fucking up the
class.}
\usage{
  \method{[}{chain} (x, i)

  \method{[}{chain} (x, i, value) <- value
}
\arguments{
  \item{x}{object to be extracted}

  \item{i}{index}

  \item{value}{value to be assigned}
}
\description{
  Need this to keep the extraction operator from fucking up
  the class.
}
\author{
  Grady Weyenberg
}

