\name{hist.chain}
\alias{hist.chain}
\title{Create histograms of variables in chain.}
\usage{
  \method{hist}{chain} (x, discard = 0.1, ...)
}
\arguments{
  \item{x}{chain object}

  \item{discard}{number or proportion of samples to discard
  as burnin.}

  \item{...}{passed to hist calls.}
}
\description{
  Create histograms of variables in chain.
}
\author{
  Grady Weyenberg
}

