\name{ld.wt.crossover}
\alias{ld.wt.crossover}
\title{associated crossover density function}
\usage{
  ld.wt.crossover(new, old)
}
\arguments{
  \item{old}{old population matrix of parents}

  \item{new}{newly proposed population matrix of parents}
}
\description{
  evaluation of density depends on the two rows that were
  randomly selected for crossover.
}
\author{
  Edward A. Roualdes
}

