\name{print.chain}
\alias{print.chain}
\title{You probably don't really want to see the entire list of states.}
\usage{
  \method{print}{chain} (x, ...)
}
\arguments{
  \item{x}{chain to summarize}

  \item{...}{passed to print}
}
\description{
  You probably don't really want to see the entire list of
  states.
}

