\name{prune}
\alias{prune}
\title{Subset each element of a chain.}
\usage{
  prune(chain, ...)
}
\arguments{
  \item{chain}{chain to operate on.}

  \item{...}{passed to the `[` function for each element in
  chain.}
}
\value{
  a new chain
}
\description{
  Subset each element of a chain.
}
\author{
  Grady Weyenberg
}

