\name{weighted.crossover}
\alias{weighted.crossover}
\title{Proposes a density weighted k=1 crossover.}
\usage{
  weighted.crossover(state)
}
\arguments{
  \item{state}{population matrix of parents}
}
\value{
  a new state
}
\description{
  Proposes a density weighted k=1 crossover.
}
\author{
  Edward A. Roualdes
}

