\name{[.chain}
\alias{[.chain}
\title{Need this to keep the extraction operator from fucking up the
class.}
\usage{
  \method{[}{chain} (x, i)
}
\arguments{
  \item{x}{object from which to extract elements}

  \item{i}{index}
}
\description{
  Need this to keep the extraction operator from fucking up
  the class.
}
\author{
  Grady Weyenberg
}

